refactor(Elife): move query into separate module
