+++ src/themes/elife/test/dataProvider.test.ts
@@ -2,10 +2,6 @@
 import * as dataProvider from '../lib/dataProvider'
 import Mock = jest.Mock
 
-interface Response {
-  ok: boolean
-  json: () => Promise<unknown>
-}
 const body = document.body
 
 const resetDom = (): void => {
@@ -14,66 +10,6 @@ const resetDom = (): void => {
 
 describe('data Provider ', () => {
   afterEach(resetDom)
-
-  describe('query', () => {
-    describe('successfully querying a valid article id', () => {
-      it('does not throw', async () => {
-        const fetchMock = (): Promise<Response> =>
-          Promise.resolve({
-            ok: true,
-            json: () => Promise.resolve(),
-          })
-        await expect(
-          // @ts-expect-error
-          dataProvider.query('validArticleId', fetchMock)
-        ).resolves.not.toThrow()
-      })
-
-      it('it exposes the url of the article PDF', async () => {
-        const fetchMock = (): Promise<Response> =>
-          Promise.resolve({
-            ok: true,
-            json: () => Promise.resolve({ pdf: 'path-to-the.pdf' }),
-          })
-        // @ts-expect-error
-        await expect(dataProvider.query('someId', fetchMock)).resolves.toEqual({
-          articleData: { pdf: 'path-to-the.pdf' },
-          ok: true,
-        })
-      })
-
-      it('it exposes the url of the figures PDF', async () => {
-        const fetchMock = (): Promise<Response> =>
-          Promise.resolve({
-            ok: true,
-            json: () =>
-              Promise.resolve({
-                figuresPdf: 'path-to-the-figures.pdf',
-              }),
-          })
-        // @ts-expect-error
-        await expect(dataProvider.query('someId', fetchMock)).resolves.toEqual({
-          articleData: { figuresPdf: 'path-to-the-figures.pdf' },
-          ok: true,
-        })
-      })
-    })
-
-    describe('being given an invalid article id', () => {
-      it('throws an ReferenceError', async () => {
-        const fetchMock = (): Promise<Response> =>
-          Promise.resolve({ ok: false, json: () => Promise.resolve() })
-        await expect(
-          // @ts-expect-error
-          dataProvider.query('invalidArticleId', fetchMock)
-        ).rejects.toThrow(
-          new Error(
-            `There was a problem getting article data for invalidArticleId`
-          )
-        )
-      })
-    })
-  })
 
   describe("getting PDF URLs for an article's id", () => {
     let articleId: string
